requirements.txt
requests>=2.28.1
pyproj>=3.4.0
Pillow>=9.2.0
rasterio>=1.3.0 